pyproject.toml
[tool.poetry]
name = "rest-pizza-delivery"
version = "0.1.0"
description = ""
authors = ["AlexandrSmolyachkovGH <alexandr.smolyachkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
pydantic = "^2.10.4"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
sqlalchemy-utils = "^0.41.2"
uvicorn = "^0.34.0"
alembic = "^1.14.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
